requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.1
PyGithub==2.2.0
Markdown==3.5.2
python-frontmatter==1.1.0
Pillow==10.2.0 